Snmp - added doc and refactor of the tests
