+++ docs/modules/ROOT/pages/reference/extensions/snmp.adoc
@@ -39,3 +39,13 @@ Please refer to the above link for usage and configuration details.
 ifeval::[{doc-show-user-guide-link} == true]
 Check the xref:user-guide/index.adoc[User guide] for more information about writing Camel Quarkus applications.
 endif::[]
+
+[id="extensions-snmp-camel-quarkus-limitations"]
+== Camel Quarkus limitations
+
+SNMP extension uses a different protocol implementation (`org.snmp4j.snmp4j`) than plain Camel 3.18. (https://snmp.app/dist/release/org/snmp4j/snmp4j/2.8.15/snmp4j-2.8.15-distribution.zip[link] for the sources)
+
+Implementation (`org.snmp4j.snmp4j`) will be used in Camel 4 and Camel Quarkus 3.
+A new provider brings more stability and brings fixes to lots of vulnerabilities.
+The configuration of the SNMP extension stays the same.
+
