+++ extensions-jvm/snmp/runtime/src/main/java/org/apache/camel/quarkus/component/snm/graal/SnmpRecorder.java
@@ -16,44 +16,17 @@
  */
 package org.apache.camel.quarkus.component.snm.graal;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
-import java.util.Vector;
-import java.util.concurrent.TimeoutException;
 
 import io.quarkus.runtime.RuntimeValue;
 import io.quarkus.runtime.annotations.Recorder;
 import org.apache.camel.Endpoint;
-import org.apache.camel.Exchange;
 import org.apache.camel.Producer;
+import org.apache.camel.component.snmp.QuarkusSnmpProducer;
+import org.apache.camel.component.snmp.QuarkusSnmpTrapProducer;
 import org.apache.camel.component.snmp.SnmpActionType;
 import org.apache.camel.component.snmp.SnmpComponent;
 import org.apache.camel.component.snmp.SnmpEndpoint;
-import org.apache.camel.component.snmp.SnmpMessage;
-import org.apache.camel.component.snmp.SnmpProducer;
-import org.apache.camel.component.snmp.SnmpTrapProducer;
-import org.apache.camel.support.DefaultProducer;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.snmp4j.CommunityTarget;
-import org.snmp4j.PDU;
-import org.snmp4j.Snmp;
-import org.snmp4j.TransportMapping;
-import org.snmp4j.event.ResponseEvent;
-import org.snmp4j.mp.MPv3;
-import org.snmp4j.mp.SnmpConstants;
-import org.snmp4j.security.SecurityModels;
-import org.snmp4j.security.SecurityProtocols;
-import org.snmp4j.security.USM;
-import org.snmp4j.smi.Address;
-import org.snmp4j.smi.GenericAddress;
-import org.snmp4j.smi.Integer32;
-import org.snmp4j.smi.OID;
-import org.snmp4j.smi.OctetString;
-import org.snmp4j.smi.VariableBinding;
-import org.snmp4j.transport.DefaultTcpTransportMapping;
-import org.snmp4j.transport.DefaultUdpTransportMapping;
 
 @Recorder
 public class SnmpRecorder {
@@ -97,256 +70,4 @@ public class SnmpRecorder {
             }
         }
     }
-
-    //code  from Camel 3.18.6
-    static class QuarkusSnmpProducer extends DefaultProducer {
-
-        private static final Logger LOG = LoggerFactory.getLogger(SnmpProducer.class);
-
-        private SnmpEndpoint endpoint;
-
-        private Address targetAddress;
-        private USM usm;
-        private CommunityTarget target;
-        private SnmpActionType actionType;
-        private PDU pdu;
-
-        public QuarkusSnmpProducer(SnmpEndpoint endpoint, SnmpActionType actionType) {
-            super(endpoint);
-            this.endpoint = endpoint;
-            this.actionType = actionType;
-        }
-
-        @Override
-        protected void doStart() throws Exception {
-            super.doStart();
-
-            this.targetAddress = GenericAddress.parse(this.endpoint.getAddress());
-            LOG.debug("targetAddress: {}", targetAddress);
-
-            this.usm = new USM(SecurityProtocols.getInstance(), new OctetString(MPv3.createLocalEngineID()), 0);
-            try {
-                SecurityModels.getInstance().addSecurityModel(this.usm);
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-
-            // setting up target
-            this.target = new CommunityTarget();
-            this.target.setCommunity(new OctetString(endpoint.getSnmpCommunity()));
-            this.target.setAddress(this.targetAddress);
-            this.target.setRetries(this.endpoint.getRetries());
-            this.target.setTimeout(this.endpoint.getTimeout());
-            this.target.setVersion(this.endpoint.getSnmpVersion());
-
-            this.pdu = new PDU();
-            // in here,only POLL do set the oids
-            if (this.actionType == SnmpActionType.POLL) {
-                for (OID oid : this.endpoint.getOids()) {
-                    this.pdu.add(new VariableBinding(oid));
-                }
-            }
-            this.pdu.setErrorIndex(0);
-            this.pdu.setErrorStatus(0);
-            this.pdu.setMaxRepetitions(0);
-            // support POLL and GET_NEXT
-            if (this.actionType == SnmpActionType.GET_NEXT) {
-                this.pdu.setType(PDU.GETNEXT);
-            } else {
-                this.pdu.setType(PDU.GET);
-            }
-        }
-
-        @Override
-        protected void doStop() throws Exception {
-            super.doStop();
-
-            try {
-                SecurityModels.getInstance().removeSecurityModel(new Integer32(this.usm.getID()));
-            } finally {
-                this.targetAddress = null;
-                this.usm = null;
-                this.target = null;
-                this.pdu = null;
-            }
-        }
-
-        @Override
-        public void process(final Exchange exchange) throws Exception {
-            // load connection data only if the endpoint is enabled
-            Snmp snmp = null;
-            TransportMapping<? extends Address> transport = null;
-
-            try {
-                LOG.debug("Starting SNMP producer on {}", this.endpoint.getAddress());
-
-                // either tcp or udp
-                if ("tcp".equals(this.endpoint.getProtocol())) {
-                    transport = new DefaultTcpTransportMapping();
-                } else if ("udp".equals(this.endpoint.getProtocol())) {
-                    transport = new DefaultUdpTransportMapping();
-                } else {
-                    throw new IllegalArgumentException("Unknown protocol: " + this.endpoint.getProtocol());
-                }
-
-                snmp = new Snmp(transport);
-
-                LOG.debug("Snmp: i am sending");
-
-                snmp.listen();
-
-                if (this.actionType == SnmpActionType.GET_NEXT) {
-                    // snmp walk
-                    List<SnmpMessage> smLst = new ArrayList<>();
-                    for (OID oid : this.endpoint.getOids()) {
-                        this.pdu.clear();
-                        this.pdu.add(new VariableBinding(oid));
-
-                        boolean matched = true;
-                        while (matched) {
-                            ResponseEvent responseEvent = snmp.send(this.pdu, this.target);
-                            if (responseEvent == null || responseEvent.getResponse() == null) {
-                                break;
-                            }
-                            PDU response = responseEvent.getResponse();
-                            String nextOid = null;
-                            Vector<? extends VariableBinding> variableBindings = response.getVariableBindings();
-                            for (int i = 0; i < variableBindings.size(); i++) {
-                                VariableBinding variableBinding = variableBindings.elementAt(i);
-                                nextOid = variableBinding.getOid().toDottedString();
-                                if (!nextOid.startsWith(oid.toDottedString())) {
-                                    matched = false;
-                                    break;
-                                }
-                            }
-                            if (!matched) {
-                                break;
-                            }
-                            this.pdu.clear();
-                            pdu.add(new VariableBinding(new OID(nextOid)));
-                            smLst.add(new SnmpMessage(getEndpoint().getCamelContext(), response));
-                        }
-                    }
-                    exchange.getIn().setBody(smLst);
-                } else {
-                    // snmp get
-                    ResponseEvent responseEvent = snmp.send(this.pdu, this.target);
-
-                    LOG.debug("Snmp: sended");
-
-                    if (responseEvent.getResponse() != null) {
-                        exchange.getIn().setBody(new SnmpMessage(getEndpoint().getCamelContext(), responseEvent.getResponse()));
-                    } else {
-                        throw new TimeoutException("SNMP Producer Timeout");
-                    }
-                }
-            } finally {
-                try {
-                    transport.close();
-                } catch (Exception e) {
-                }
-                try {
-                    snmp.close();
-                } catch (Exception e) {
-                }
-            }
-        } //end process
-    }
-
-    //code  from Camel 3.18.6
-    static class QuarkusSnmpTrapProducer extends DefaultProducer {
-
-        private static final Logger LOG = LoggerFactory.getLogger(SnmpTrapProducer.class);
-
-        private SnmpEndpoint endpoint;
-
-        private Address targetAddress;
-        private USM usm;
-        private CommunityTarget target;
-
-        public QuarkusSnmpTrapProducer(SnmpEndpoint endpoint) {
-            super(endpoint);
-            this.endpoint = endpoint;
-        }
-
-        @Override
-        protected void doStart() throws Exception {
-            super.doStart();
-
-            this.targetAddress = GenericAddress.parse(this.endpoint.getAddress());
-            LOG.debug("targetAddress: {}", targetAddress);
-
-            this.usm = new USM(SecurityProtocols.getInstance(), new OctetString(MPv3.createLocalEngineID()), 0);
-            SecurityModels.getInstance().addSecurityModel(this.usm);
-
-            // setting up target
-            this.target = new CommunityTarget();
-            this.target.setCommunity(new OctetString(endpoint.getSnmpCommunity()));
-            this.target.setAddress(this.targetAddress);
-            this.target.setRetries(this.endpoint.getRetries());
-            this.target.setTimeout(this.endpoint.getTimeout());
-            this.target.setVersion(this.endpoint.getSnmpVersion());
-        }
-
-        @Override
-        protected void doStop() throws Exception {
-            super.doStop();
-
-            try {
-                SecurityModels.getInstance().removeSecurityModel(new Integer32(this.usm.getID()));
-            } finally {
-                this.targetAddress = null;
-                this.usm = null;
-                this.target = null;
-            }
-        }
-
-        @Override
-        public void process(final Exchange exchange) throws Exception {
-            // load connection data only if the endpoint is enabled
-            Snmp snmp = null;
-            TransportMapping<? extends Address> transport = null;
-
-            try {
-                LOG.debug("Starting SNMP Trap producer on {}", this.endpoint.getAddress());
-
-                // either tcp or udp
-                if ("tcp".equals(this.endpoint.getProtocol())) {
-                    transport = new DefaultTcpTransportMapping();
-                } else if ("udp".equals(this.endpoint.getProtocol())) {
-                    transport = new DefaultUdpTransportMapping();
-                } else {
-                    throw new IllegalArgumentException("Unknown protocol: " + this.endpoint.getProtocol());
-                }
-
-                snmp = new Snmp(transport);
-
-                LOG.debug("SnmpTrap: getting pdu from body");
-                PDU trap = exchange.getIn().getBody(PDU.class);
-
-                trap.setErrorIndex(0);
-                trap.setErrorStatus(0);
-                if (this.endpoint.getSnmpVersion() == SnmpConstants.version1) {
-                    trap.setType(PDU.V1TRAP);
-                } else {
-                    trap.setType(PDU.TRAP);
-                    trap.setMaxRepetitions(0);
-                }
-
-                LOG.debug("SnmpTrap: sending");
-                snmp.send(trap, this.target);
-                LOG.debug("SnmpTrap: sent");
-            } finally {
-                try {
-                    transport.close();
-                } catch (Exception e) {
-                }
-                try {
-                    snmp.close();
-                } catch (Exception e) {
-                }
-            }
-        } //end process
-    }
-
 }
